เพิ่มเงื่อนไข ลบ Plan

+++ CIM.API/Controllers/CacheController.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
 using CIM.BusinessLogic.Interfaces;
+using CIM.DAL.Interfaces;
 using CIM.Model;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -19,16 +21,19 @@ namespace CIM.API.Controllers
         private IActiveProductionPlanService _activeProductionPlanService;
         private IMachineService _machineService;
         private IResponseCacheService _responseCacheService;
+        private IProductionPlanRepository _productionPlanRepository;
 
         public CacheController(
             IActiveProductionPlanService activeProductionPlanService,
             IMachineService machineService,
-            IResponseCacheService responseCacheService
+            IResponseCacheService responseCacheService,
+            IProductionPlanRepository productionPlanRepository
             )
         {
             _activeProductionPlanService = activeProductionPlanService;
             _machineService = machineService;
             _responseCacheService = responseCacheService;
+            _productionPlanRepository = productionPlanRepository;
         }
 
         [HttpGet]
@@ -91,11 +96,28 @@ namespace CIM.API.Controllers
             var planList = planIds.Split(',');
             try
             {
-                foreach (var planId in planList)
-                {   
-                    await _activeProductionPlanService.RemoveCached(planId);
+                var data = _productionPlanRepository.Where(x => x.PlanId == planIds).Select(x => x.StatusId).ToArray();
+                if (data[0] == 3)
+                {
+                    foreach (var planId in planList)
+                    {
+                        await _activeProductionPlanService.RemoveCached(planId);
+                    }
+                    return "OK";
                 }
-                return "OK";
+                else if (data[0] == 2)
+                {
+                    return "Plan Running";
+                }
+                else if (data[0] == 1)
+                {
+                    return "Plan not Started";
+                }
+                else
+                {
+                    return "Without Plan";
+                }
+                
             }
             catch (Exception ex)
             {
